requirements.txt
streamlit
google-generativeai
beautifulsoup4
requests
python-dotenv
pillow
google-auth-oauthlib
google-api-python-client
